% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{abkineticsModel}
\alias{abkineticsModel}
\title{Define antibody kinetics model}
\format{
A list with the following columns:
\describe{
\item{names}{Unique identifier for each exposure type.}
\item{model}{Description of the model using the \code{makeModel} and \code{addAbkineticsModel} functions.}
\item{prior}{Description of the prior distribution using the \code{add_par_df} function.}
}
}
\description{
This data format represents information about the antibody kinetics
}
\details{
Add more data about this model here.
}
\examples{
# Example usage. This describes the antibody kinetics for a SARS-CoV-2 delta wave using IgG data. First define the antibody kinetics function:

infSerumKinetics <- function(titre_est, timeSince, pars) {
   a <- pars[1] 
   b <- pars[2] 
   c <- pars[3]
   if (timeSince < 14) {
       titre_est <- titre_est + (log(exp(a) + exp(c)) * (timeSince) / 14);
   } else {
       titre_est <- titre_est + (log(exp(a) * exp(-b/10 * (timeSince - 14))) + exp(c));
   }
   titre_est
}

# Now define the antibody kinetics function in the format required for the rjmc package:

abkineticsModel <- list(
   names = c("delta"),
   model = makeModel(
           addAbkineticsModel("IgG", "delta", TRUE, c("a_d", "b_d", "c_d"), infSerumKinetics),
       ),
   prior = bind_rows(
       add_par_df("a_d", -2, 2, "norm",  0, 1), # ab kinetics
       add_par_df("b_d", 0, 1, "norm",  0.3, 0.05), # ab kinetics
       add_par_df("c_d", 0, 4, "unif", 0,  4), # ab kinetics
   )


}
\seealso{
\code{\link{makeModel}}, \code{\link{addAbkineticsModel}}, \code{\link{add_par_df}},  for related functions.
}
\author{
David Hodgson
}
\keyword{antibody}
\keyword{data}
\keyword{format}
\keyword{kinetics}
\keyword{model}
