% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_front.R
\name{add_par_df}
\alias{add_par_df}
\title{add_par_df}
\usage{
add_par_df(...)
}
\arguments{
\item{...}{The list as described.}
}
\description{
This function adds information about the prior distributions of the parameter in the serojump model. It takes a list of the form:
\itemize{
\item par_name: The name of the parameter as as string.
\item lb: The numerical lower bound of the parameter.
\item ub: The numerical upper bound of the parameter.
\item dist: The distribution describing the prior distribution from standard R disributions (e.g. unif, norm)
\item dist_par1: The first argument of the prior distribution,
\item dist_par2: The second argument of the prior distribution, (NA if only one parameter is taken like in exp)
}
}
