% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model.R
\name{addAbkineticsModelHier}
\alias{addAbkineticsModelHier}
\title{addAbkineticsModelHier}
\usage{
addAbkineticsModelHier(
  id,
  biomarker,
  exposureType,
  pars,
  parsHier,
  dataHier,
  funcForm
)
}
\arguments{
\item{id}{The name of the biomarker.}

\item{biomarker}{The name of the biomarker.}

\item{exposureType}{The name of the exposure type.}

\item{pars}{The parameters of the model.}

\item{parsHier}{The parameters of the Hierarchical model.}

\item{dataHier}{The data for the Hierarchical model.}

\item{funcForm}{The antibody kinetics function.}
}
\value{
A list with the biomarker name, the exposure name, whether the exposure is inferred, the parameters and the antibody kinetics function.
}
\description{
This function adds an antibody kinetics model to the model definition.
}
