% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model.R
\name{addCopModel}
\alias{addCopModel}
\title{addCopModel}
\usage{
addCopModel(biomarker, exposureType, pars, funcForm, logLikelihood)
}
\arguments{
\item{biomarker}{The name of the biomarker.}

\item{exposureType}{The name of the exposure type.}

\item{pars}{The parameters of the model.}

\item{funcForm}{The functional form of the observational model.}

\item{logLikelihood}{The log likelihood function.}
}
\value{
A list with the biomarker name, the exposure name, the parameters and the log likelihood function.
}
\description{
This function adds a cop model to the model definition.
}
