% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_front.R
\name{addExposurePrior}
\alias{addExposurePrior}
\title{This function adds the exposure prior to the rjmc model.}
\usage{
addExposurePrior(model_type, data_t, exp_prior, type = NULL)
}
\arguments{
\item{model_type}{rjmc_model created from createModelRJCMCFull()}

\item{data_t}{data included in the model created from generate_data_t()}

\item{exp_prior}{a data.frame describing either a functional or empirical prior}

\item{type}{define a type of exposure prior, \code{func} for functional and 'empirical' for empirical. If blank it will assume a uniform distribution between 1 and T.}
}
\value{
returns a rjmc model
}
\description{
This function takes two types of priors. Either a functional prior or an empirical prior.
}
\seealso{
createModelRJCMCFull() and generate_data_t()
}
