% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model.R
\name{addObservationalModel}
\alias{addObservationalModel}
\title{addObservationalModel}
\usage{
addObservationalModel(biomarker, pars, logLikelihood)
}
\arguments{
\item{biomarker}{The name of the biomarker.}

\item{pars}{The parameters of the model.}

\item{logLikelihood}{The log likelihood function.}
}
\value{
A list with the biomarker name, the parameters and the log likelihood function.
}
\description{
This function adds an observational model to the model definition.
}
