% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_front.R
\name{check_exposures_times}
\alias{check_exposures_times}
\title{check_exposures_times}
\usage{
check_exposures_times(
  data_sero,
  known_exp,
  exposure_types,
  fitted_exp,
  exposurePriorTime = NULL,
  exposurePriorTimeType = NULL
)
}
\arguments{
\item{data_sero}{A data.frame with columns id, time and \code{biomarkers}}

\item{known_exp}{A data.frame with columns id, time and exposure_type}

\item{exposure_types}{A vector of the exposure types}

\item{fitted_exp}{A string of the exposure type that is fitted}

\item{exposurePriorTime}{A data.frame describing either a functional or empirical prior}

\item{exposurePriorTimeType}{define a type of exposure prior, \code{func} for functional and 'empirical' for empirical. If blank it will assume a uniform distribution between 1 and T.}
}
\value{
A warning if there are values outside this range
}
\description{
check_exposures_times
}
\details{
This function checks that the data_sero has exposure times within the range the bleeds taken for an individual
}
\note{
This function can be used to check the data_sero before running the serojump model.
}
