% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_front.R
\name{check_input_data}
\alias{check_input_data}
\title{check_input_data}
\usage{
check_input_data(data_sero, known_inf)
}
\arguments{
\item{data_sero}{A data.frame with columns id, time and IgG}

\item{known_inf}{A data.frame with columns id, time and exposure_type}
}
\value{
A warning if there are single entries
}
\description{
check_input_data
}
\details{
This function checks the data_sero input for the serojump model.
}
