% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_front.R
\name{check_sero_no_single_entries}
\alias{check_sero_no_single_entries}
\title{check_no_single_entries}
\usage{
check_sero_no_single_entries(data_sero)
}
\arguments{
\item{data_sero}{A data.frame with columns id, time and IgG}
}
\value{
A warning if there are single entries
}
\description{
check_no_single_entries
}
\details{
This function checks that the data_sero has more than one observation per person.
}
\note{
This function can be used to check the data_sero before running the serojump model.
}
