% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model.R
\name{copLogLikelihood}
\alias{copLogLikelihood}
\title{Define the Log Likelihood function for the COP model}
\arguments{
\item{inf_status}{The infection status of an individual binary (0 or 1).}

\item{esttitreExp}{The model-estimated titre value at the time of exposure}

\item{pars}{The fitted parameters needed to calculate the log likelihood. These are defined in the \code{prior} entry of the copModel list.}
}
\value{
A function that returns the log likelihood value for the COP model.
}
\description{
This function is defined by the user and describes the COP likelihood of the data given the model.
}
\details{
Add information here.
}
\examples{
# Example usage:
copLogLikelihood <- function(inf_status, esttitreExp, params) {
   # COP parameters
   beta0 <- params[1]
   beta1 <- params[2]
   p <- 1.0 / (1.0 + exp(- (beta0 + beta1 * esttitreExp) ) )
   ll <- inf_status * log(p) + (1 - inf_status) * log(1 - p)
   ll
}


}
\author{
Your Name
}
\keyword{data}
\keyword{export}
\keyword{format}
\keyword{function}
