% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{copModel}
\alias{copModel}
\title{Define correlate of protection model}
\format{
A list with the following columns:
\describe{
\item{names}{Unique identifier for each biomarker.}
\item{model}{Description of the model using the \code{makeModel} and \code{addObservationalModel} functions.}
\item{prior}{Description of the prior distribution using the \code{add_par_df} function.}
}
}
\description{
This data format represents information about the model for the correlate of protection (probability of infection given exposure) given the titre of a biomarker for a predefined exposure type
}
\details{
Add more data about this model here.
}
\examples{
# Example usage. This describes the observation model for a SARS-CoV-2 delta wave using IgG data. First define the COP log likelihood function, for example if the COP is a logistic function of the titre of the biomarker IgG (see \ref{obsLogLikelihood} for format):

copLogLikelihood <- function(ll, inf_status, esttitreExp, params) {
   # COP parameters
   beta0 <- params[1]
   beta1 <- params[2]
   p <- 1.0 / (1.0 + exp(- (beta0 + beta1 * esttitreExp) ) )
   ll <- inf_status * log(p) + (1 - inf_status) * log(1 - p)
   ll
}


}
\seealso{
\code{\link{makeModel}}, \code{\link{addObservationalModel}}, \code{\link{add_par_df}},  for related functions.
}
\author{
David Hodgson
}
\keyword{COP}
\keyword{data}
\keyword{format}
\keyword{model}
