% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_front.R
\name{createModelRJCMCFull}
\alias{createModelRJCMCFull}
\title{This function create the rjmc model.}
\usage{
createModelRJCMCFull(ab_ll, par_tab)
}
\arguments{
\item{ab_ll}{the likelihood functions define by a function called \code{evaluateLogLikelihood}}

\item{par_tab}{a data.frame describing the prior distributions}

\item{cop_func}{a function for the correlate of protection}
}
\value{
returns a rjmc model
}
\description{
By taking the defined prior distributions, this creates the rjmc model including the support, names of parameters, prior pdf, prior smapling function, the intial conditions for the exposure timings and adds the correlate of protection.
}
\seealso{
createModelRJCMCFull() and generate_data_t()
}
