% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model.R
\name{createSeroJumpModel}
\alias{createSeroJumpModel}
\title{createSeroJumpModel}
\usage{
createSeroJumpModel(
  data_sero,
  data_known,
  biomarkers,
  exposureTypes,
  exposureFitted,
  observationalModel,
  abkineticsModel,
  exposurePriorTime = NULL,
  exposurePriorTimeType = NULL,
  exposurePriorPop = NULL,
  known_exp_bool = NULL,
  seed = -1
)
}
\arguments{
\item{data_sero}{The serology data.}

\item{data_known}{The known exposure data.}

\item{biomarkers}{A vector of the biomarkers}

\item{exposureTypes}{A vector of the exposure types}

\item{exposureFitted}{The exposure type that is fitted}

\item{observationalModel}{The observational model}

\item{abkineticsModel}{The antibody kinetics model}

\item{exposurePriorTime}{The prior time of the exposure}

\item{exposurePriorTimeType}{The type of the exposure prior}

\item{exposurePriorPop}{The population prior of the exposure}

\item{known_exp_bool}{Boolean describing if the exposures are known in the model}
}
\value{
A list with the data and the model.
}
\description{
This function creates a model for the serology jump model.
}
