% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{data_known_exposures}
\alias{data_known_exposures}
\title{Input the known exposures}
\format{
A list with the following columns:
\describe{
\item{id}{Unique identifier for each biomarker.}
\item{time}{Time at which the known exposure occurred}
\item{exposure_type}{Key for the type of exposure experienced}
}

A list with the following columns:
\describe{
\item{id}{Unique identifier for each biomarker.}
\item{time}{Time at which the known exposure occurred}
\item{exposure_type}{Key for the type of exposure experienced}
}
}
\description{
This data format represents the serological data used by the rjmcmc model

This data format represents the serological data used by the rjmcmc model
}
\details{
Add more data about this model here.

Add more data about this model here.
}
\examples{
# Example usage. For example if you have some information on known vaccinatoins, you can define:

data_titre <- data.frame(
     id = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10),
     time = c(10, 4, 100, 40, 20, 23, 26, 70, 40, 10),
     exposure_type = rep("vax", 10)
)


# Example usage. For example if you have some information on known vaccinatoins, you can define:

data_titre <- data.frame(
     id = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10),
     time = c(10, 4, 100, 40, 20, 23, 26, 70, 40, 10),
     exposure_type = rep("vax", 10)
)


}
\seealso{
Add related functions

Add related functions
}
\author{
David Hodgson
}
\keyword{data}
\keyword{data_known_exposures}
\keyword{format}
