% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{data_titre_ex}
\alias{data_titre_ex}
\title{Input the serological data}
\format{
A list with the following columns:
\describe{
\item{id}{Unique identifier for each biomarker.}
\item{time}{Description of the model using the \code{makeModel} and \code{addObservationalModel} functions.}
\item{titre}{Description of the prior distribution using the \code{addPrior} function.}
\item{biomarker}{Description of the prior distribution using the \code{addPrior} function.}
}
}
\description{
This data format represents the serological data used by the rjmcmc model
}
\details{
Add more data about this model here.
}
\examples{
# Example usage. For example we can make some summy data for IgG SARS-CoV-2

data_titre <- data.frame(
     id = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10),
    time = c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0),
   titre = c(1.5, 1.6, 1.7, 1.8, 1.9, 2.0, 2.1, 2.2, 2.3, 2.4),
     biomarker = rep("IgG", 10)
)


}
\seealso{
Add related functions
}
\author{
David Hodgson
}
\keyword{data}
\keyword{data_titre}
\keyword{format}
