% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_front.R
\name{generate_data_t}
\alias{generate_data_t}
\title{This function creates data_t, which is a list of data to use in the rjmc model}
\usage{
generate_data_t(data_titre_model, data_inf_model = NULL, known_exp = NULL)
}
\arguments{
\item{data_titre_model}{is a data.frame with column headings \link{'id', 'titre', 'time'} where 'id' is the individual, 'titre' is the titre value and 'time' is the time of the titre. If not defined correctly it will throw an error.}

\item{data_inf_model}{is a data.frame with column headings \link{'id', 'inf_time'} where 'id' is the individual and 'inf_time' is the time of infection. If not defined it will assume no known infection}

\item{known_exp}{add a known exposure vector to the model. If not defined it will assume no known exposure}
}
\value{
returns a list of data
}
\description{
This function requires a \code{data_titre_model} data.frame from the user, which is the observed titre data, and a \code{data_inf_model} data.frame from the user, which is the observed infection data. The function will then create a list of data to use in the rjmc model.
}
