% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model.R
\name{makeModel}
\alias{makeModel}
\title{makeModel}
\usage{
makeModel(...)
}
\arguments{
\item{...}{The models to include in the definition.}
}
\value{
A list with the models.
}
\description{
This function creates a model list for observationalModel, copModel, and abKineticsModel.
}
