% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model.R
\name{obsLogLikelihood}
\alias{obsLogLikelihood}
\title{Define the Log Likelihood function for the observational model}
\arguments{
\item{titre_val}{The titre value from the data}

\item{titre_est}{The model-estimate titre value.}

\item{pars}{The fitted parameters needed to calculate the log likelihood. These are defined in the \code{prior} entry of the \ref{observationalModel} list.}
}
\value{
A functoion that returns the log likelihood value for the COP model.
}
\description{
This function is defined by the user and describes the observational likelihood of the data given the model.
}
\details{
Add information here.
}
\examples{
# Example usage:
obsLogLikelihood = function(titre_val, titre_est, pars) {
   if (titre_val <= log10(40)) {
       ll <- pcauchy(log10(40), titre_est, pars[1], log.p = TRUE)
   } else {
       ll <- dcauchy(titre_val, titre_est, pars[1], log = TRUE)
   }
   ll
}


}
\author{
Your Name
}
\keyword{data}
\keyword{export}
\keyword{format}
\keyword{function}
