% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{plotPostFigs}
\alias{plotPostFigs}
\title{plotPostFigs}
\usage{
plotPostFigs(model_summary, save_info)
}
\arguments{
\item{model_summary}{Fit from the serojump model}

\item{save_info}{Information about saving the figures}
}
\description{
plotPostFigs
}
