% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_sim.R
\name{plotPostFigsSim}
\alias{plotPostFigsSim}
\title{plotPostFigs}
\usage{
plotPostFigsSim(model_summary, sim_model, sim_res, save_info)
}
\arguments{
\item{model_summary}{Fit from the serojump model}

\item{sim_model}{The model description from serosim}

\item{sim_res}{The fitted simulated model from serosim}

\item{save_info}{Information about saving the figures}
}
\description{
plotPostFigs
}
