% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity.R
\name{plotPriorInfection}
\alias{plotPriorInfection}
\title{A plotting function for the prior on the infection rate}
\usage{
plotPriorInfection(seroModel)
}
\arguments{
\item{seroModel}{A serological model object}
}
\value{
A ggplot object with the plotted data
}
\description{
This function takes a serological model object and plots the prior distribution of the exposure rate both on the population-level and individual-level.
}
