% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity.R
\name{plotPriorPredictive}
\alias{plotPriorPredictive}
\title{A plotting function for the prior predictive distribution of the antibody kinetics}
\usage{
plotPriorPredictive(seroModel)
}
\arguments{
\item{seroModel}{A serological model object}
}
\value{
A ggplot object with the plotted data
}
\description{
This function takes a serological model object and plots the prior predictive distribution of the antibody kinetics. The plot gives an overview of the prior predictive distribution of the antibody kinetics
}
