% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity.R
\name{plotSero}
\alias{plotSero}
\title{A plotting function for the serological data}
\usage{
plotSero(seroModel, N_max = 100)
}
\arguments{
\item{seroModel}{A serological model object}

\item{N_max}{Maximum number of individuals to plot}
}
\value{
A ggplot object with the plotted data
}
\description{
This function takes a serological model object and plots the serological data. The plot gives an overview of the data across individuals
}
