% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_sim_inf.R
\name{postprocessFigs_simInf}
\alias{postprocessFigs_simInf}
\title{Postprocess the posteriors of the simulation study}
\usage{
postprocessFigs_simInf(modelname, modelname_sim, obs_er, n_chains)
}
\arguments{
\item{modelname}{The name of the model}

\item{modelname_sim}{The name of the simulation model}

\item{obs_er}{The observation error}

\item{n_chains}{The number of chains}
}
\description{
Postprocess the posteriors of the simulation study
}
