% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmc_sero.R
\name{rjmc_sero_func}
\alias{rjmc_sero_func}
\title{Create a model}
\usage{
rjmc_sero_func(model, data, settings, par = NULL)
}
\arguments{
\item{model}{The model to run ptmc.}

\item{data}{Data used in the calibration process}

\item{settings}{settings}

\item{par}{NULL}
}
\value{
Returns a list with the fist element being the mcmc samples formatting for analysis and plottig with the CODA package. The second is the log posterior value at each time step.
}
\description{
Create a model
}
