% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model.R
\name{runRJMCMC}
\alias{runRJMCMC}
\alias{runInfRJMCMC}
\title{run the RJMCMC algorithm}
\usage{
runRJMCMC(seroModel, settings, filename, modelname)

runInfRJMCMC(seroModel, settings, filename, modelname, priorPred = TRUE)
}
\arguments{
\item{seroModel}{The seroModel previously defined.}

\item{settings}{Settings used for the calibration}

\item{filename}{Filepath of where the outputs are saved (outputs/fits/filename)}

\item{modelname}{Name of the model outputs (in outputs/fits/filename)}
}
\value{
A list with the posterior samples, the model and the data.

A list with the posterior samples, the model and the data.
}
\description{
This function runs the RJMCMC algorithm given a defined seroModel, settings and filepaths for output

This function runs the RJMCMC algorithm given a defined seroModel, settings and filepaths for output
}
