% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_diag.R
\name{seroJumpPostDaig}
\alias{seroJumpPostDaig}
\title{Postprocess the posteriors of the simulation study}
\usage{
seroJumpPostDaig(fitfull, filepathfig)
}
\arguments{
\item{modelname}{The nams of the model}

\item{modelname_sim}{The name of the simulation model}

\item{obs_er}{The observation error}

\item{n_chains}{The number of chains}
}
\description{
Postprocess the posteriors of the simulation study
}
