% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model.R
\name{createSeroJumpModel}
\alias{createSeroJumpModel}
\title{createSeroJumpModel}
\usage{
createSeroJumpModel(
  data_sero,
  data_known,
  modeldefinition,
  known_exp_bool = NULL
)
}
\arguments{
\item{data_sero}{The serology data.}

\item{data_known}{The known exposure data.}

\item{modeldefinition}{The model definition.}
}
\value{
A list with the data and the model.
}
\description{
This function creates a model for the serology jump model.
}
