% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_diag.R
\name{plotMCMCDiagnosis}
\alias{plotMCMCDiagnosis}
\title{Postprocess the posteriors of the simulation study}
\usage{
plotMCMCDiagnosis(model_summary, save_info)
}
\arguments{
\item{model_summary}{The fitted model from serojump}

\item{save_info}{Information on how to save the diagnosis figures}
}
\description{
Postprocess the posteriors of the simulation study
}
